Prime some tests to avoid not found errors
Its possible that, in some apis, due to dynamic table construction
a call might result in a transient not found error (the evp method
lookup being an example).  Retrying fixes that, but since this is a
performance test, avoid the possibility by calling the do function for a
given test once before starting the multithreaded test.  This guarantees
our lookups will be found consistently.  We should however come up with
a mechanism to ensure table population prior to multithreaded use