Add a release validation script
Currently, part of release process involves (causally) confirming that
the release artifacts pass gpg/sha1/sha256sum checks.  It would be nice
if we could automate that a bit.  This script allows a user to
automatically download draft release artifacts and confirm that their
signatures/sha1sum/sha256 sum match those of the included artifact files

It also introduces the optional feature to do a per-file level
validation of the release using an SBOM file.  If we choose to include
SBOMS in our releases (see
https://github.com/openssl/openssl/pull/29131), this script will for
each file in the archive:
1) confirm that a node in the SBOM file exists for it
2) confirm that the sha256sum of the file in the archive matches that of
   the sum included in the SBOM
3) confirm that the sha256sum of the corresponding file from the git
   tree matches that of the sum included in the SBOM

By confirming these three elements we can have a greater degree of
confidence that our automated release pipeline did not alter any of our
source files from the initial source that we obtained from git.